CS 501 Binary Tree Project
Inputs 20 strings into a binary tree. Prints out the inorder, preorder and postorder traversals indicating the level of each node as it's printed out. Also prints the number of leaves in the tree, the number of nodes with one child, number of nodes with two children, maximum and minimum leaf levels, height of tree, and whether or not the tree is balanced.